Issue #104: make service and action methods.
There were both properties previously. Moved to methods for
a more dynamic use, where the service can change according to
a modifying parameter.

Removed __construct() from Response, as we are only passing
an array in as data now to initialise, and not a body object.
This was a legacy left-over, and good to finally remove it.